551th Kattis Solution. Apply Ceasar Cipher iteratively and process the result.
This one was a bit tricky to get the answer just right so I use a lot of asserts in the code. The function okForBoris() runs on the assertion that "fef" is false, this is because "fef" is one of those edge cases. It contains exactly the double amount of consonants as compared to vowels. Overall the problem was tricky with regards to edge cases hence all the asserts to make sure that every function does what it is supossed to do.