653th Kattis. Bit of an ugly solution.
This solution is what we must call a two-sack solution. There is code so ugly you can only sit with it provided that you place a sack over your computer. Then there are pieces of code so hideous you too must wear a sack, lest the computer's sack falls off.